services/ml/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
numpy==2.2.6
openai==1.101.0
langgraph==0.6.6
httpx==0.28.1
faiss-cpu==1.12.0
sentence-transformers==5.1.0
pytest==8.4.1
